# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Compare repos",
        "description": "Compare two repos",
        "contact": {
            "email": "dudekpavlo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/api/comparator/find": {
            "post": {
                "tags": [
                    "comparator"
                ],
                "summary": "Get info about repos",
                "operationId": "find",
                "parameters": [
                    {
                        "name": "firstPackage",
                        "in": "query",
                        "description": "First package",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "secondPackage",
                        "in": "query",
                        "description": "Second package",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repos data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Repo data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}